Moved ixSuggestions factory to Factory
